MS1/requirements.txt
flask
pika
psycopg2-binary
argon2-cffi